Capitalize helper in a route and a controller
DASHED HELPER NAMESIn Ember versions before 1.13, helpers needed to have a dash in their name or they needed to be explicitly registered. This was mostly for performance reasons but with the Glimmer engine that restriction went away. If you work on a legacy Ember app, this is nevertheless useful to know and can save you some headscratching.