Begining of animating star-component
Our first thought would be to replace #each with #liquid-each and there is actually an open issue in the liquid-fire repository to add a liquid-each helper, but it does not yet exist. We achieve the same effect, though, with wrapping the #each in a liquid-bind

We wrapped the #each in a liquid-bind and looped through the variable yielded back by it, liquidStars, in the #each. Whenever stars changes, the liquid-bind will observe that change and pass it to the transition map. Because stars changes whenever the song's rating changes, this fits our needs.