Continuing stars animation, creat transition func
this.inHelper matches a transition that was observed by the helper that we pass to it, in this case, liquid-bind. As we currently have only one instance of that helper, this will work, though it might be prudent to add a css class and then use this.hasClass(...), just as we did for the band's description textarea.