Undo star-rating animations
The "animation markup" (the liquid-bind helper) in the template of the star- rating component is disquieting. After all, we should be able to use the rating stars with- or without animations in the same project.The component's template should be independent of other libraries in the application, and we violated this reasonable assumption by including a liquid-fire helper in the template.

This could be solved by "escaping forward". We'll extract the star-rating widget into an Ember addon and provide it with a block form invocation so that it can be freely customized.